[GOHW-3]: fix sku

+++ loms/migrations/00001_add_cart_item_table.sql
@@ -3,7 +3,7 @@
 CREATE TABLE cart_item
 (
   user_id bigint,
-  sku int,
+  sku bigint,
   name text,
   price int,
   count int
